Pre-release snapshot of 4.3
This is intended to be a bug fix release which is expected to be released
in December.

There are some code infrastructure changes, which I belive to be in
preparation for a later feature release (4.4) that might happen in about
6 months (which could easily change).

There are lots of fixes for better handling of bad data, only two of
which have been backported into Fedora's 4.2 packages.