Keep setup output out of test.log
It looks like the command that gets run covers more than the test script.
We need to keep that other output out of the output we search for
failure. The includes the dnf step, since it can have non fatal failures.
It also looks like the dnf step might need kernel-modules-extra for
mounting squashfs file systems.