4.5 has been released
his is a bare bones update to test the release. There were a lot of
changes and 2 new sym lynced names. The programs help has info
about this. But the man pages are missing a lot and still need to
be updated. Potentially command line changes could break composes,
though I am not expecting that.