Pick up a few upstream fixes.
These are pretty minor and I wasn't planning on building anywhere but
rawhide for the time being. I think there is a good chance of additional
minor updates within a month or two. In particular, it would make sense
to tag a minor release. But I'd like to get a little wider testing of
the changes so far, because we found some issues right after the last
tagged release, because we didn't do enough testing.