Get some wider testing before a new squashfs-tools point release
Phillip is going to issue a new point release soon. Adding man pages is
the main thing to still be done. There are a few changes since our last
bug fix update. I want to get a little wider testing of the changes
before Phillip tags the new release. I'd rather not find something
just after the releae like what happened with 4.5.

I'm not going to push this to released versions of Fedora. Once the
new point release is out, I expect to eventually push that to released
versions of Fedora.