Get lmza man pages improvements
There is now better documentation in the man page and help for
the deprecated lzma support when it is built with that enabled.

There is still a bit more stuff expected before the 4.5.1 release.
I don't expect updates to be pushed back to released versions of
Fedora until that release occurs or something unexpected happens.

The point of getting these updates into rawhide is to make it less
likely that a regression gets into the 4.5.1 release.